# day01-cpp - part 2, 73x faster