--- a/day01-chronal-calibration/main.cpp
+++ b/day01-chronal-calibration/main.cpp
@@ -1,7 +1,9 @@
 #include <iostream>
 #include <string>
-#include <set>
+#include <unordered_set>
 #include <vector>
+
+#include <chrono>
 
 int main(int argc, char** argv)
 {
@@ -16,34 +18,36 @@ int main(int argc, char** argv)
         input.push_back(atoi(line.c_str()));
     }
 
+
     // Calculates puzzle 1
     int total = 0;
-    for (auto var : input)
+    for (const auto& var : input)
     {
         total += var;
     }
-
     std::cout << total << '\n';
 
 
     // Calculates puzzle 2;
-    std::set<int> duplicates;
+    std::unordered_set<int> duplicates;
     int current = 0;
     bool found = false;
+
+    // auto begin = std::chrono::steady_clock::now();
     while (!found)
     {
-        for (auto var : input)
+        for (const auto& var : input)
         {
             current += var;
-
-            if (duplicates.find(current) != duplicates.end()) {
-                std::cout << current << '\n';
+            if (!duplicates.insert(current).second) {
                 found = true;
                 break;                
             }
-            
-            duplicates.insert(current);
         }
     }
+    // auto end = std::chrono::steady_clock::now();
+       // std::cout << "T2 = " << std::chrono::duration_cast<std::chrono::microseconds>(end - begin).count() << "us\n";
+
+    std::cout << current << '\n';
     return 0;
 }