--- a/day01-chronal-calibration/main.cpp
+++ b/day01-chronal-calibration/main.cpp
@@ -29,10 +29,15 @@ int main(int argc, char** argv)
 
 
     // Calculates puzzle 2;
-    std::unordered_set<int> duplicates{};
-    constexpr int n262144 = 2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2;
-    duplicates.reserve(n262144);
-    int current = 0;
+    //constexpr int n262144 = 2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2;
+    //std::unordered_set<int> duplicates{};
+    //duplicates.reserve(n262144);
+    
+    constexpr int32_t n32768  = 2*2*2*2*2*2*2*2*2*2*2*2*2*2*2;
+    constexpr int32_t n524288 = 2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2;
+
+    int32_t bit_field[n32768] = {};
+    int32_t current = 0;
     bool found = false;
 
     auto begin = std::chrono::steady_clock::now();
@@ -41,15 +46,19 @@ int main(int argc, char** argv)
         for (const auto& var : input)
         {
             current += var;
-            if (!duplicates.insert(current).second) {
+            const int32_t field_index = (current + n524288) / 32;
+            const int32_t bit_index = (current + n524288) % 32;
+
+            if (bit_field[field_index] & (0b1 << bit_index)) {
                 found = true;
-                break;                
+                break;
             }
+            bit_field[field_index] |= (0b1 << bit_index);
         }
     }
     auto end = std::chrono::steady_clock::now();
     std::cout << "T2 = " << std::chrono::duration_cast<std::chrono::microseconds>(end - begin).count() << "us\n";
 
-    std::cout << current << " " << duplicates.size() <<'\n';
+    std::cout << current <<'\n';
     return 0;
 }