This was an experiment to see how much performance for day01part2 could improve from the naive implementation. I intend to only do this optimization for day01, as it is generally better if the code is easier to read, than being really fast. At least in this project.

**Notes**
- Testing on MBP2015
- Starting at 22ms for part 2 of day01
- 22ms -> 15ms - using `unordered_map` instead of `map`
- 15ms -> 11ms - using unordered_map::reserve, to allocate memory needed up-front.
- CPP solution  on-par with rust performance
- 11ms -> 0.35ms !!! - using bit-field to store visited numbers.
- Dangerous, only supports integers in range [-250k, 250k] now, but it is enough to solve for the given input data. Range could easily increase, if needed.
- 0.35ms -> 0.30ms by `>> 5` instead of  `/ 32`