@tholok97 2018-12-16
`bit_field` takes a punch at readability, but performance > readability in this project I guess :yum: