@Arxcis 2018-12-16
> `bit_field` takes a punch at readability, but performance > readability in this project I guess 😋

I just did this for fun. I don't have any plans to do this on any of the other tasks. But it was a fun experiement to see how far I could go with this.