@Arxcis 2018-12-17
> `bit_field` takes a punch at readability, but performance > readability in this project I guess yum

I actually am tempted to make it a requirement to not use bit-wise operations in any of the solutions (unless required specifically by the given task), just to make sure benchmarks reflect everyday usage of the languages.