@tholok97 2018-12-17
It is cool to see how far you can push performance as well though.

If there was a good way to support alternate solutions to a given day and language that could be a solution. Could get messy though.